+++ authors/apps/authentication/backends.py
@@ -18,22 +18,16 @@ class JWTAuthentication(authentication.BaseAuthentication):
         auth_header = authentication.get_authorization_header(request).split()
         auth_header_prefix = self.authentication_header_prefix.lower()
 
-        if not auth_header:
-            return None
+        if len(auth_header) == 2:
 
-        if len(auth_header) == 1:
-            return None
+            prefix = auth_header[0].decode('utf-8')
+            token = auth_header[1].decode('utf-8')
 
-        elif len(auth_header) > 2:
-            return None
+            if prefix.lower() != auth_header_prefix:
+                return None
 
-        prefix = auth_header[0].decode('utf-8')
-        token = auth_header[1].decode('utf-8')
-
-        if prefix.lower() != auth_header_prefix:
-            return None
-
-        return self._authenticate_credentials(request, token)
+            return self._authenticate_credentials(request, token)
+        return None
 
     def _authenticate_credentials(self, request, token):
         """
