+++ authors/apps/core/utils.py
@@ -21,31 +21,20 @@ def validate_registration(data):
             }
         )
 
-    if not username:
-        raise serializers.ValidationError(
-            {
-                'input fields':
-                'Please input a username'
-            }
-        )
-    if len(username) < 4:
-        raise serializers.ValidationError(
-            {
-                'username':
-                'Please the username should be at'
-                'least 4 characters'
-            }
-        )
+    validate_username_isnot_empty(username)
 
-    if not re.match(r"^[A-Za-z]+[\d\w_]+", username):
-        raise serializers.ValidationError(
+    validate_username_length(username)
 
-            {
-                'username':
-                'Username should start with letters'
-            }
-        )
+    validate_username(username)
 
+    validate_email(email)
+
+    valid_password(password)
+
+    return{"username": username, "email": email, "password": password}
+
+
+def validate_email(email):
     if not re.search(r"([\w\.-]+)@([\w\.-]+)(\.[\w\.]+$)", email):
         raise serializers.ValidationError(
             {
@@ -55,8 +44,37 @@ def validate_registration(data):
             }
         )
 
-    valid_password(password)
-    return{"username": username, "email": email, "password": password}
+
+def validate_username_length(username):
+    if len(username) < 4:
+        raise serializers.ValidationError(
+            {
+                'username':
+                'Please the username should be at'
+                'least 4 characters'
+            }
+        )
+
+
+def validate_username_isnot_empty(username):
+    if not username:
+        raise serializers.ValidationError(
+            {
+                'input fields':
+                'Please input a username'
+            }
+        )
+
+
+def validate_username(username):
+    if not re.match(r"^[A-Za-z]+[\d\w_]+", username):
+        raise serializers.ValidationError(
+
+            {
+                'username':
+                'Username should start with letters'
+            }
+        )
 
 
 def valid_password(password):
