+++ authors/apps/renderers.py
@@ -12,11 +12,22 @@ class MainRenderer(JSONRenderer):
     object_label = 'object'
 
     def render(self, data, accepted_media_type=None, renderer_context=None):
+        self.is_data_list(data)
+        self.is_data_dictionary(data)
+
+    def is_data_list(self, data):
+        """
+        Checks whether data type is a list.
+        """
         if type(data) != ReturnList:
             errors = data.get('errors', None)
             if errors is not None:
                 return super(MainRenderer, self).render(data)
 
+    def is_data_dictionary(self, data):
+        """
+        Checks whether data type is a dictionary.
+        """
         if type(data) == ReturnDict:
             return json.dumps({
                 self.object_label: data
